+++ components/common/PageLayout/components/DocumentHistory/DocumentHistory.tsx
@@ -1,30 +1,31 @@
 import styled from '@emotion/styled';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
-import { memo, useEffect, useMemo, useState } from 'react';
+import { memo, useCallback, useEffect, useState } from 'react';
 
 import { useMembers } from 'hooks/useMembers';
 import type { PageMeta } from 'lib/pages';
-import { getRelativeTime } from 'lib/utilities/getRelativeTime';
+import { getRelativeTimeInThePast } from 'lib/utilities/dates';
 
 const StyledTypography = styled(Typography)`
   color: ${({ theme }) => theme.palette.grey[300]};
 `;
 
 function DocumentHistory ({ page }: { page: PageMeta }) {
-  const updatedAtMemo = useMemo(() => getRelativeTime(new Date(page.updatedAt)), [page.updatedAt]);
-  const createdAtMemo = useMemo(() => getRelativeTime(new Date(page.createdAt)), [page.createdAt]);
-
-  const [{ updatedAt, createdAt }, setTime] = useState({ updatedAt: updatedAtMemo, createdAt: createdAtMemo });
+  const updatedAtCb = useCallback(() => ({
+    createdAt: getRelativeTimeInThePast(new Date(page.createdAt)),
+    updatedAt: getRelativeTimeInThePast(new Date(page.updatedAt))
+  }), [page.updatedAt, page.updatedAt]);
+  const [{ updatedAt, createdAt }, setTime] = useState(updatedAtCb());
 
   const { members } = useMembers();
 
-  const createdBy = members.find(member => member.id === page.createdBy)?.username ?? 'unknown';
+  const createdBy = members.find(member => member.id === page.createdBy)?.username ?? 'Unknown user';
   const updatedBy = members.find(member => member.id === page.updatedBy)?.username ?? createdBy;
 
   useEffect(() => {
-    setTime((prevState) => ({ ...prevState, updatedAt: updatedAtMemo }));
-  }, [updatedAtMemo]);
+    setTime(updatedAtCb());
+  }, [updatedAtCb]);
 
   return (
     <Tooltip
@@ -32,7 +33,7 @@ function DocumentHistory ({ page }: { page: PageMeta }) {
       enterDelay={0}
       arrow={false}
       placement='bottom-start'
-      onOpen={() => setTime({ updatedAt: getRelativeTime(new Date(page.updatedAt)), createdAt: getRelativeTime(new Date(page.createdAt)) })}
+      onOpen={() => setTime(updatedAtCb())}
       title={(
         <>
           <StyledTypography variant='caption'>Edited by</StyledTypography> {updatedBy} <StyledTypography variant='caption'>{updatedAt}</StyledTypography>
